Updated the backend webhook project.
Earlier we also sent the user-id along with token, amount.

Now we are sending the token, webhook-url, amount. The rest of the details will
be queried, once the transfer is approved/rejected.

Also, added a common 'type' as package, to store all commonly used types and interfaces.
Here, it'll be used on both webhook and dummy-bank.